requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
boto3==1.42.66
botocore==1.42.66
certifi==2026.1.4
cffi==2.0.0
cryptography==46.0.5
dynamodb-encryption-sdk==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.1.0
numpy==2.4.3
ollama==0.6.1
pandas==3.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
